Starting work on async settings
Attempt fixing tests, starting to think about the best way to go about this hmm

Replace old HTTP2 test server and conn tests with new versions.

Fix integration tests.
